add startswith?, endswith?, File class

+++ vm/src/array_obj.h
@@ -7,7 +7,7 @@ extern "C" {
 #include "value.h"
 typedef array(struct value) array_obj_data;
 struct array_obj {
-    size_t refs;
+    uint32_t refs;
     array_obj_data data;
 };
 void array_obj_init(struct array_obj *);
